# src/Domain/Enum/PermissionEnum.php
<?php

namespace App\Domain\Enum;

enum PermissionEnum: string
{
    // Survey
    case SurveyView = 'survey.view';
    case SurveyComplete = 'survey.complete';
    case SurveyCreate = 'survey.create';
    case SurveyUpdate = 'survey.update';
    case SurveyDelete = 'survey.delete';
    case SurveyViewResult = 'survey.view_result';
    case SurveyViewAll = 'survey.view_all';

    // User
    case UserCreate = 'user.create';
    case UserView = 'user.view';
}